fix: gnome-keyring-daemon environment vars
Something changed about the way gnome-keyring-daemon prints its
variables causing two variables to glob together with the quotes.

I'm guessing the fact that there is one called
$GNOME_KEYRING_CONTROL that prints before the $SSH_AUTH_SOCK
is the crux of the issue.

Anyway, this fixes that behavior by not quoting the string before the
export command so we get two separate variables exported.

Also, adds some nice sauce so we get the right behavior when we log in
via ssh or su in a zsh shell.